+++ R/forest.netcomparison.R
@@ -208,6 +208,7 @@ forest.netcomparison <- function(x,
     suppressWarnings(metagen(TE, seTE, data = dat, sm = x$x$sm,
                              studlab = comparison, backtransf = backtransf,
                              level = x$level,
+                             method.tau = "DL",
                              warn = FALSE))
   ##
   forest(m1,